migration to sql server

+++ BestMovies.Net/BestMovies.Api/Models/Movie.cs
@@ -2,7 +2,7 @@
 
 public class Movie
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Title { get; set; } = null!;
-    public short? Year { get; set; }
+    public int? Year { get; set; }
 }
