+++ BestMovies.Net/BestMovies.Api/Models/Person.cs
@@ -2,7 +2,7 @@
 
 public class Person
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Name { get; set; } = null!;
-    public short? Birth { get; set; }
+    public int? Birth { get; set; }
 }
