+++ BestMovies.Net/BestMovies.Api/Models/Star.cs
@@ -2,8 +2,8 @@
 
 public class Star
 {
-    public long MovieId { get; set; }
-    public long PersonId { get; set; }
+    public int MovieId { get; set; }
+    public int PersonId { get; set; }
 
     public virtual Movie Movie { get; set; } = null!;
     public virtual Person Person { get; set; } = null!;
